@@ -1,7 +1,7 @@
 {
   "name": "nino",
   "description": "Moderation bot for Discord's hack week entry.",
-  "version": "0.8.2",
+  "version": "0.9.0",
   "homepage": "https://nino.augu.dev",
   "license": "MIT",
   "repository": "https://github.com/NinoDiscord/Nino",
@@ -15,7 +15,9 @@
     "main": "node dist/index.js",
     "lint": "eslint src --ext .ts --fix",
     "test": "jest --config jest.config.js --forceExit",
-    "dev": "nodemon dist/index.js"
+    "dev": "nodemon dist/index.js",
+    "docker": "docker build -t nino:latest .",
+    "main:docker": "docker create --name Nino nino"
   },
   "maintainers": [
     "August (Chris)",
@@ -27,8 +29,9 @@
     "ohlookitsderpy"
   ],
   "dependencies": {
-    "@augu/immutable": "0.3.1",
-    "@sentry/node": "5.19.2",
+    "@augu/immutable": "0.3.2",
+    "@augu/sysinfo": "0.2.1",
+    "@sentry/node": "5.20.1",
     "common-tags": "1.8.0",
     "eris": "0.13.3",
     "inversify": "5.0.1",
@@ -36,26 +39,27 @@
     "ioredis": "4.17.3",
     "js-yaml": "3.14.0",
     "leeks.js": "0.0.8",
-    "mongoose": "5.9.24",
+    "mongoose": "5.9.27",
     "ms": "2.1.2",
     "reflect-metadata": "0.1.13",
     "wumpfetch": "0.3.1"
   },
   "devDependencies": {
     "@types/common-tags": "1.8.0",
-    "@types/ioredis": "4.17.2",
-    "@types/jest": "26.0.4",
+    "@types/ioredis": "4.17.3",
+    "@types/jest": "26.0.8",
     "@types/js-yaml": "3.12.4",
-    "@types/mongoose": "5.7.32",
+    "@types/mongoose": "5.7.34",
     "@types/ms": "0.7.31",
     "@types/node": "13.13.5",
-    "@typescript-eslint/eslint-plugin": "3.6.1",
-    "@typescript-eslint/parser": "3.6.1",
-    "eslint": "7.4.0",
-    "jest": "26.1.0",
+    "@typescript-eslint/eslint-plugin": "3.8.0",
+    "@typescript-eslint/parser": "3.8.0",
+    "eslint": "7.6.0",
+    "jest": "26.2.2",
     "jest-mock-extended": "1.0.9",
     "nodemon": "2.0.4",
-    "ts-jest": "26.1.2",
+    "ps-list": "7.2.0",
+    "ts-jest": "26.1.4",
     "typescript": "3.9.7"
   }
 }
