@@ -3,12 +3,23 @@ import { stripIndents } from 'common-tags';
 import { TYPES } from '../../types';
 import Context from '../../structures/Context';
 import Command from '../../structures/Command';
-import Bot from '../../structures/Bot';
+import Bot, { Config } from '../../structures/Bot';
+import CommandManager from '../../structures/managers/CommandManager';
+import Eris from 'eris';
+import { createEmptyEmbed } from '../../util/EmbedUtils';
+import { lazyInject } from '../../inversify.config';
 
 @injectable()
 export default class HelpCommand extends Command {
-  constructor(@inject(TYPES.Bot) client: Bot) {
-    super(client, {
+  @lazyInject(TYPES.CommandManager)
+  private commandManager!: CommandManager;
+
+  constructor(
+    @inject(TYPES.Bot) bot: Bot, 
+    @inject(TYPES.Client) private client: Eris.Client,
+    @inject(TYPES.Config) private config: Config
+  ) {
+    super(bot, {
       name: 'help',
       description: (bot) => `Gives a list of ${bot.client.user ? bot.client.user.username : 'Nino'}'s commands or shows documentation on a specific command`,
       usage: '[command]',
@@ -16,8 +27,8 @@ export default class HelpCommand extends Command {
     });
   }
 
-  getAllCategories() {
-    const commands = this.bot.manager.commands.filter(x => !x.hidden);
+  private getAllCategories() {
+    const commands = this.commandManager.commands.filter(x => !x.hidden);
     const categories: { [x: string]: Command[]; } = {};
     for (const command of commands) {
       if (!(command.category in categories)) categories[command.category] = [];
@@ -27,48 +38,53 @@ export default class HelpCommand extends Command {
     return categories;
   }
 
-  async run(ctx: Context) {
+  private async sendCommandList(ctx: Context) {
     const settings = await ctx.getSettings();
     const categories = this.getAllCategories();
+    const commands = this.commandManager.commands;
+    const botUser = this.client.user;
 
+    const title = ctx.translate('commands.generic.help.embed.title', {
+      username: `${botUser.username}#${botUser.discriminator}`
+    });
+
+    const description = ctx.translate('commands.generic.help.embed.description', {
+      website: 'https://nino.augu.dev',
+      commands: commands.size
+    });
+
+    const footer = ctx.translate('commands.generic.help.embed.footer', {
+      prefix: settings ? settings.prefix : this.config.discord.prefix
+    });
+
+    const embed = createEmptyEmbed()
+      .setTitle(title)
+      .setDescription(description)
+      .setFooter(footer);
+
+    for (const category in categories) {
+      const localized = ctx.translate(`commands.generic.help.categories.${category.toLowerCase()}`);
+      embed.addField(`${localized} [${categories[category].length}]`, categories[category].map(s => `\`${s.name}\``).join(', '));
+    }
+
+    return ctx.embed(embed.build());
+  }
+
+  async run(ctx: Context) {
     if (!ctx.args.has(0)) {
-      const title = ctx.translate('commands.generic.help.embed.title', {
-        username: `${ctx.bot.client.user.username}#${ctx.bot.client.user.discriminator}`
-      });
+      return this.sendCommandList(ctx);
+    }
 
-      const description = ctx.translate('commands.generic.help.embed.description', {
-        website: 'https://nino.augu.dev',
-        commands: this.bot.manager.commands.size
-      });
+    const arg = ctx.args.get(0);
+    const command = this.commandManager.getCommand(arg);
 
-      const footer = ctx.translate('commands.generic.help.embed.footer', {
-        prefix: settings ? settings.prefix : this.bot.config.discord.prefix
-      });
-
-      const embed = ctx.bot.getEmbed()
-        .setTitle(title)
-        .setDescription(description)
-        .setFooter(footer);
-
-      for (const category in categories) {
-        const localized = ctx.translate(`commands.generic.help.categories.${category.toLowerCase()}`);
-        embed.addField(`${localized} [${categories[category].length}]`, categories[category].map(s => `\`${s.name}\``).join(', '));
-      }
-
-      return ctx.embed(embed.build());
-    } else {
-      const arg = ctx.args.get(0);
-      const command = ctx.bot.manager.getCommand(arg);
-
-      const notFound = ctx.translate('commands.generic.help.notFound', {
+    if (!command) {
+      return ctx.sendTranslate('commands.generic.help.notFound', {
         command: arg
       });
-
-      if (!command) return ctx.send(notFound);
-      else {
-        const embed = await command.help(ctx);
-        return ctx.embed(embed);
-      }
+    } else {
+      const embed = await command.help(ctx);
+      return ctx.embed(embed);
     }
   }
 }
