@@ -4,11 +4,16 @@ import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
 import Bot from '../../structures/Bot';
+import { createEmptyEmbed } from '../../util/EmbedUtils';
+import Eris from 'eris';
 
 @injectable()
 export default class InviteCommand extends Command {
-  constructor(@inject(TYPES.Bot) client: Bot) {
-    super(client, {
+  constructor(
+    @inject(TYPES.Bot) bot: Bot,
+    @inject(TYPES.Client) private client: Eris.Client
+  ) {
+    super(bot, {
       name: 'invite',
       description: 'Gives you the invite for the bot.',
       aliases: ['inv'],
@@ -17,10 +22,10 @@ export default class InviteCommand extends Command {
   }
 
   async run(ctx: Context) {
-    const embed = this.bot.getEmbed()
+    const embed = createEmptyEmbed()
       .setTitle(ctx.translate('commands.generic.invite.title'))
       .setDescription(ctx.translate('commands.generic.invite.description', {
-        invite: `<https://discordapp.com/oauth2/authorize?client_id=${ctx.me.id}&scope=bot>`,
+        invite: `<https://discordapp.com/oauth2/authorize?client_id=${this.client.user.id}&scope=bot>`,
         server: 'https://discord.gg/7TtMP2n'
       }))
       .build();
