@@ -1,18 +1,31 @@
-import { humanize, formatSize } from '../../util';
+import { humanize, formatSize, commitHash } from '../../util';
 import { injectable, inject } from 'inversify';
 import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
 import Bot from '../../structures/Bot';
+import sys from '@augu/sysinfo';
+import { Client } from 'eris';
+import CommandStatisticsManager from '../../structures/managers/CommandStatisticsManager';
+import DatabaseManager from '../../structures/managers/DatabaseManager';
+import { createEmptyEmbed } from '../../util/EmbedUtils';
+import CommandManager from '../../structures/managers/CommandManager';
+import { lazyInject } from '../../inversify.config';
 
 const pkg = require('../../../package.json');
 
 @injectable()
 export default class StatisticsCommand extends Command {
+  @lazyInject(TYPES.CommandManager)
+  private commandManager!: CommandManager;
+
   constructor(
-    @inject(TYPES.Bot) client: Bot
+    @inject(TYPES.Bot) bot: Bot,
+    @inject(TYPES.Client) private client: Client,
+    @inject(TYPES.CommandStatisticsManager) private commandStatisticsManager: CommandStatisticsManager,
+    @inject(TYPES.DatabaseManager) private databaseManager: DatabaseManager
   ) {
-    super(client, {
+    super(bot, {
       name: 'statistics',
       description: 'Gives you the bot\'s statistics',
       aliases: ['stats', 'info', 'bot', 'botinfo']
@@ -20,31 +33,41 @@ export default class StatisticsCommand extends Command {
   }
 
   async run(ctx: Context) {
-    const { command, uses } = this.bot.statistics.getCommandUsages();
-    const users = this.bot.client.guilds.reduce((a, b) => a + b.memberCount, 0).toLocaleString();
-    const channels = Object.keys(this.bot.client.channelGuildMap).length;
-    const shardPing = this.bot.client.shards.reduce((a, b) => a + b.latency, 0);
-    const connection = await this.bot.database.admin.ping();
+    const { command, uses } = this.commandStatisticsManager.getCommandUsages();
+    const guilds = this.client.guilds;
+    const shards = this.client.shards;
+    const users = guilds.reduce((a, b) => a + b.memberCount, 0).toLocaleString();
+    const channels = Object.keys(this.client.channelGuildMap).length;
+    const shardPing = shards.reduce((a, b) => a + b.latency, 0);
+    const connection = await this.databaseManager.admin.ping();
     const memoryUsage = formatSize(process.memoryUsage().rss);
+    const botUser = this.client.user;
 
-    const embed = this.bot.getEmbed()
-      .setTitle(ctx.translate('commands.generic.statistics.title', { username: `${this.bot.client.user.username}#${this.bot.client.user.discriminator}` }))
+    const embed = createEmptyEmbed()
+      .setAuthor(ctx.translate('commands.generic.statistics.title', { 
+        username: `${botUser.username}#${botUser.discriminator}`,
+        version: `v${pkg.version} / ${commitHash}`
+      }), 'https://nino.augu.dev', botUser.dynamicAvatarURL('png', 1024))
       .setDescription(ctx.translate('commands.generic.statistics.description', {
-        guilds: this.bot.client.guilds.size.toLocaleString(),
+        guilds: guilds.size.toLocaleString(),
         users,
         channels,
-        current: ctx.guild ? ctx.guild.shard.id : 0,
-        total: this.bot.client.shards.size,
-        latency: shardPing,
+        'shards.alive': shards.filter(s => s.status === 'ready').length,
+        'shards.total': shards.size,
+        'shards.latency': shardPing,
         uptime: humanize(Math.round(process.uptime()) * 1000),
-        commands: this.bot.manager.commands.size,
-        messages: this.bot.statistics.messagesSeen.toLocaleString(),
-        executed: this.bot.statistics.commandsExecuted.toLocaleString(),
-        name: command,
-        executions: uses,
-        connected: connection.ok === 1 ? ctx.translate('global.online') : ctx.translate('global.offline'),
-        memoryUsage,
-        version: pkg.version
+        commands: this.commandManager.commands.size,
+        messagesSeen: this.commandStatisticsManager.messagesSeen.toLocaleString(),
+        commandsExecuted: this.commandStatisticsManager.commandsExecuted.toLocaleString(),
+        'mostUsed.command': command,
+        'mostUsed.executed': uses,
+        'memory.rss': memoryUsage,
+        'memory.heap': formatSize(process.memoryUsage().heapUsed),
+        'sys.os': sys.getPlatform(),
+        'sys.cpu': sys.getCPUUsage(),
+        'sys.free': formatSize(sys.getFreeMemory()),
+        'sys.total': formatSize(sys.getTotalMemory()),
+        connection
       }))
       .build();
 
