@@ -4,13 +4,15 @@ import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
 import Bot from '../../structures/Bot';
+import { Client } from 'eris';
 
 @injectable()
 export default class UptimeCommand extends Command {
   constructor(
-    @inject(TYPES.Bot) client: Bot
+    @inject(TYPES.Bot) bot: Bot,
+    @inject(TYPES.Client) private client: Client
   ) {
-    super(client, {
+    super(bot, {
       name: 'uptime',
       description: 'Gives you the uptime for the bot.',
       aliases: ['up']
@@ -19,7 +21,7 @@ export default class UptimeCommand extends Command {
 
   async run(ctx: Context) {
     const pUptime = Math.round(process.uptime()) * 1000;
-    const bUptime = Date.now() - this.bot.client.startTime;
+    const bUptime = Date.now() - this.client.startTime;
 
     return ctx.sendTranslate('commands.generic.uptime', {
       connection: humanize(bUptime),
