@@ -3,6 +3,7 @@ import { Guild, TextChannel } from 'eris';
 import { TYPES } from '../types';
 import Client from '../structures/Bot';
 import Event from '../structures/Event';
+import { createEmptyEmbed } from '../util/EmbedUtils';
 
 @injectable()
 export default class GuildLeftEvent extends Event {
@@ -22,7 +23,7 @@ export default class GuildLeftEvent extends Event {
     const channel = await this.bot.client.getRESTChannel('529593466729267200');
     if (channel.type === 0) {
       const chan = (channel as TextChannel);
-      const embed = this.bot.getEmbed()
+      const embed = createEmptyEmbed()
         .setAuthor(`| Left ${guild.name} (${guild.id})`, undefined, this.bot.client.user.dynamicAvatarURL('png', 1024))
         .setFooter(`Now at ${this.bot.client.guilds.size} Guilds`)
         .build();
