@@ -3,6 +3,8 @@ import { Punishment, PunishmentType } from './PunishmentManager';
 import { Guild } from 'eris';
 import { injectable, inject } from 'inversify';
 import { TYPES } from '../../types';
+import { bigTimeout } from '../../util';
+import ms from 'ms';
 
 /**
  * This class makes timeouts resilent, automatic unmutes and unbans will be supported even on an event of a crash.
@@ -18,14 +20,22 @@ export default class TimeoutsManager {
     this.bot = bot;
   }
 
-  /**
-   * The setTimeout function for big time values.
-   * @param func the function to execute
-   * @param time the time to excute it after
-   */
-  bigTimeout(func: (...args: any[]) => void, time: number) {
-    if (time > 0x7fffffff) setTimeout(() => this.bigTimeout(func, time - 0x7fffffff), 0x7fffffff);
-    else setTimeout(func, time);
+  private createTimeout(key: string, task: string, member: string, guild: Guild, time: number) {
+    return bigTimeout(async () => {
+      if (await this.bot.redis.hexists('timeouts', key)) {
+        try {
+          await this.bot.punishments.punish(
+            { id: member, guild },
+            new Punishment(task as PunishmentType, {
+              moderator: this.bot.client.user,
+            }),
+            '[Automod] Time\'s up!'
+          );
+        } finally {
+          await this.bot.redis.hdel('timeouts', key);
+        }
+      }
+    }, time);
   }
 
   /**
@@ -36,18 +46,10 @@ export default class TimeoutsManager {
    * @param time the amount of time before executing
    */
   async addTimeout(member: string, guild: Guild, task: 'unban' | 'unmute', time: number) {
+    this.bot.logger.info(`Added timeout: to ${task} in ${ms(time)}`);
     const key = `timeout:${task}:${guild.id}:${member}`;
-    await this.bot.redis.set(key, `${Date.now()}:${time}:${member}:${guild.id}:${task}`);
-    this.bigTimeout(async () => {
-      const exists = await this.bot.redis.exists(key);
-      if (exists) {
-        const punishment = new Punishment(task as PunishmentType, {
-          moderator: this.bot.client.user
-        });
-        await this.bot.punishments.punish({ id: member, guild }, punishment, '[Automod] Time\'s up!');
-        await this.bot.redis.del(key);
-      }
-    }, time);
+    await this.bot.redis.hset('timeouts', key, `${Date.now()}:${time}:${member}:${guild.id}:${task}`);
+    this.createTimeout(key, task, member, guild, time);
   }
 
   /**
@@ -58,38 +60,36 @@ export default class TimeoutsManager {
    */
   async cancelTimeout(member: string, guild: Guild, task: string) {
     const key = `timeout:${task}:${guild.id}:${member}`;
-    return this.bot.redis.del(key);
+    return this.bot.redis.hdel('timeouts', key);
+  }
+
+  private async resetOldTimeouts() {
+    const oldtimedates = await this.bot.redis.keys('timeout:*:*:*');
+
+    for (let timedate of oldtimedates) {
+      const value = await this.bot.redis.get(timedate);
+      await this.bot.redis.del(timedate);
+      await this.bot.redis.hset('timeouts', timedate, value!);
+    }
   }
 
   /**
    * Reapply timeouts since server went dead
    */
   async reapplyTimeouts() {
-    const timedates = await this.bot.redis.keys('timeout:*:*:*');
+    await this.resetOldTimeouts();
+    const timedates = await this.bot.redis.hkeys('timeouts');
+    
     for (let timedate of timedates) {
-      const value = await this.bot.redis.get(timedate);
+      const value = await this.bot.redis.hget('timeouts', timedate);
       const start = Number(value!.split(':')[0]);
       const amount = Number(value!.split(':')[1]);
       const member = value!.split(':')[2];
       const guild = this.bot.client.guilds.get(value!.split(':')[3]);
       const task = value!.split(':')[4];
       if (!guild) continue;
-
-      this.bigTimeout(async () => {
-        if (await this.bot.redis.exists(timedate)) {
-          try {
-            await this.bot.punishments.punish(
-              { id: member, guild },
-              new Punishment(task as PunishmentType, {
-                moderator: this.bot.client.user,
-              }),
-              '[Automod] Time\'s up!'
-            );
-          } finally {
-            await this.bot.redis.del(timedate);
-          }
-        }
-      }, start - Date.now() + amount);
+      
+      this.createTimeout(timedate, task, member, guild, start - Date.now() + amount);
     }
   }
 }
