@@ -1,35 +1,35 @@
-import model, { GuildModel } from '../../models/GuildSchema';
+import model, { UserModel } from '../../../models/UserSchema';
 import { SettingsBase as Base } from './SettingsBase';
 import { injectable, inject } from 'inversify';
-import { TYPES } from '../../types';
-import Bot from '../Bot';
+import { TYPES } from '../../../types';
+import Bot from '../../Bot';
 
 @injectable()
-export default class GuildSettings implements Base<GuildModel> {
-  public client: Bot;
+export default class UserSettingsService implements Base<UserModel> {
+  public bot: Bot;
   public model = model;
 
   constructor(@inject(TYPES.Bot) bot: Bot) {
-    this.client = bot;
+    this.bot = bot;
   }
 
   async get(id: string) {
-    const document = await this.model.findOne({ guildID: id }).exec();
+    const document = await this.model.findOne({ userID: id }).exec();
     if (!document || document === null) return null;
     return document;
   }
 
-  async getOrCreate(id: string): Promise<GuildModel> {
+  async getOrCreate(id: string) {
     let settings = await this.get(id);
     if (!settings || settings === null)
       settings = await this.create(id);
     return settings!;
   }
 
-  async create(id: string): Promise<GuildModel> {
+  async create(id: string) {
     const query = new this.model({
-      guildID: id,
-      prefix: this.client.config.discord.prefix,
+      userID: id,
+      locale: 'en_US'
     });
     await query.save();
     return query;
@@ -37,7 +37,7 @@ export default class GuildSettings implements Base<GuildModel> {
 
   remove(id: string) {
     return this.model
-      .findOne({ guildID: id })
+      .findOne({ userID: id })
       .remove()
       .exec();
   }
@@ -47,9 +47,6 @@ export default class GuildSettings implements Base<GuildModel> {
     doc: { [x: string]: any },
     cb: (error: any, raw: any) => void
   ) {
-    const search = { guildID: id };
-    // eslint-disable-next-line
-    if (!!doc.$push) search['punishments'] = { $not: { $size: 15 }};
-    return this.model.updateOne(search, doc, cb);
+    return this.model.updateOne({ userID: id }, doc, cb);
   }
 }
